Refactor some code to reuse types and utilities
